Update library with new ValueIteration and PolicyIteration implementations
